Fostrich fix temple tasks subgraph & others (#874)
* fix: use for..of instead of foreach to avoid async issues

* fix: update subgraph query data interface results

* feat: clean unused eth checker task & recheck eth balance at the end of batch liquidate bot

* feat: allow tx to try tlc batch liq execution below min eth gas config, but send discord warning
