+++ apps/automation-v2/src/subgraph/types.ts
@@ -1,11 +1,5 @@
 export type Users = {
   id: string;
-  collateral: string;
-  collateralUSD: string;
-  debt: string;
-  debtUSD: string;
-  enterTimestamp: string;
-  exitTimestamp: string;
 }[];
 
 export interface SubGraphResponse<T extends object> {
@@ -18,7 +12,7 @@ export interface SubGraphQuery {
   variables?: { [key: string]: string | number };
 }
 
-export type GetUserResponse = SubGraphResponse<{ users: Users }>;
+export type GetUserResponse = SubGraphResponse<{ tlcUsers: Users }>;
 
 export class SubgraphError extends Error {
   constructor(public message: string, public cause: Error) {
