+++ apps/automation-v2/src/tlc/batch-liquidate.ts
@@ -3,28 +3,28 @@ import { DISCORD_WEBHOOK_URL_KEY, connectDiscord } from '@/common/discord';
 import {
   TempleTaskDiscordEvent,
   TempleTaskDiscordMetadata,
+  buildDiscordMessageCheckEth,
   buildTempleTasksDiscordMessage,
   formatBigNumber,
 } from '@/common/utils';
 import { ITempleLineOfCredit, ITempleLineOfCredit__factory } from '@/typechain';
 import { EventLog } from 'ethers';
 import {
-  TaskContext,
   TaskResult,
+  TaskContext,
   taskSuccess,
   taskSuccessSilent,
 } from '@mountainpath9/overlord';
 import { subgraphRequest } from '@/subgraph/subgraph-request';
 import { GetUserResponse } from '@/subgraph/types';
 import { matchAndDecodeEvent } from '@/common/filters';
-import { buildDiscordMessageCheckEth } from '@/common/eth-auto-checker';
 
 export interface TlcBatchLiquidateConfig {
   CHAIN: Chain;
   WALLET_NAME: string;
   TLC_ADDRESS: string;
   ACC_LIQ_MAX_CHUNK_NO: number;
-  MIN_ETH_BALANCE: bigint;
+  MIN_ETH_BALANCE_WARNING: bigint;
   GAS_LIMIT: bigint;
   SUBGRAPH_URL: string;
 }
@@ -44,6 +44,8 @@ export async function batchLiquidate(
     signer
   );
 
+  const submittedAt = new Date();
+
   const chunkify = function (itr: string[], size: number) {
     const chunk: string[][] = [];
     let innerChunk: string[] = [];
@@ -58,10 +60,9 @@ export async function batchLiquidate(
     return chunk;
   };
 
-  const submittedAt = new Date();
 
   const res = await getTlcUsers(config.SUBGRAPH_URL);
-  const tlcUsers = (await res()).data?.users;
+  const tlcUsers = (await res()).data?.tlcUsers;
   // if undefined or zero users returned from subgraph, success silently
   if (!tlcUsers || tlcUsers.length === 0) return taskSuccessSilent();
   const accountsToCheck = tlcUsers.flatMap((u) => u.id);
@@ -82,9 +83,8 @@ export async function batchLiquidate(
     const tx = await tlc.batchLiquidate(accBatch, {
       gasLimit: config.GAS_LIMIT,
     });
-
     const txReceipt = await tx.wait();
-    if (!txReceipt) return taskSuccessSilent();
+    if (!txReceipt) throw Error('undefined tx receipt');
 
     // Grab the events
     const events: TempleTaskDiscordEvent[] = [];
@@ -143,32 +143,30 @@ export async function batchLiquidate(
       metadata
     );
     await discord.postMessage(message);
-
-    // Send discord alert if signer wallet doesn't have sufficient eth balance
-    const ethBalance = await provider.getBalance(walletAddress);
-    if (ethBalance < config.MIN_ETH_BALANCE) {
-      const ethBalanceMessage = await buildDiscordMessageCheckEth(
-        config.CHAIN,
-        submittedAt,
-        walletAddress,
-        ethBalance,
-        config.MIN_ETH_BALANCE
-      );
-
-      await discord.postMessage(ethBalanceMessage);
-    }
   }
 
+  // Send discord alert warning if signer wallet doesn't have sufficient eth balance
+  const ethBalance = await provider.getBalance(walletAddress);
+  if (ethBalance < config.MIN_ETH_BALANCE_WARNING) {
+    const ethBalanceMessage = await buildDiscordMessageCheckEth(
+      config.CHAIN,
+      submittedAt,
+      walletAddress,
+      ethBalance,
+      config.MIN_ETH_BALANCE_WARNING
+    );
+    await discord.postMessage(ethBalanceMessage);
+  }
   return taskSuccess();
 }
 
 const getTlcUsers = async (url: string) => {
   const resp = await subgraphRequest<GetUserResponse>(url, {
     query: `{
-        tlcUsers {
-          id
-        }
-      }`,
+      tlcUsers(where: {debt_gt: "0"}) {
+        id
+      }
+    }`,
   });
 
   return resp;
